% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auth_Requests.R
\name{ATWeb_Auth}
\alias{ATWeb_Auth}
\title{Authentication Request}
\usage{
ATWeb_Auth(username, password)
}
\arguments{
\item{username}{Username for the API}

\item{password}{Password for the API}
}
\description{
This function queries the Archway API to login.
It is used at the beginning of other requests to the API
}
