% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAPBills.R
\name{GetAPBills}
\alias{GetAPBills}
\title{GetAPBills API Call}
\usage{
GetAPBills(username, password, enterpriseID, BillStartDate, BillEndDate)
}
\arguments{
\item{username}{Username for the API}

\item{password}{Password for the API}

\item{enterpriseID}{Enterprise ID}

\item{StartDate}{Date for beginning of period in yyyy-mm-dd format}

\item{EndDate}{Date for end of period in yyyy-mm-dd format}
}
\description{
This function queries the Archway API for information on Portfolio Transactions for a given period
and returns xml
}
