% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBalanceSheet.R
\name{GetBalanceSheet}
\alias{GetBalanceSheet}
\title{GetBalanceSheet API Call}
\usage{
GetBalanceSheet(username, password, enterpriseID, AsOfDate)
}
\arguments{
\item{username}{Username for the API}

\item{password}{Password for the API}

\item{enterpriseID}{Enterprise ID}

\item{AsOfDate}{Date for end of period in yyyy-mm-dd format}
}
\description{
This function queries the Archway API for Balance Sheet data for all Entities.
}
