% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetEntities.R
\name{GetEntities}
\alias{GetEntities}
\title{GetEntities API Call}
\usage{
GetEntities(username, password, enterpriseID)
}
\arguments{
\item{username}{Username for the API}

\item{password}{Password for the API}

\item{enterpriseID}{Enterprise ID}
}
\description{
This function queries the Archway API for information on entities
}
