% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetEntities.R
\name{GetEntityClasses}
\alias{GetEntityClasses}
\title{GetEntities API Call for Class Information}
\usage{
GetEntityClasses(username, password, enterpriseID)
}
\arguments{
\item{username}{Username for the API}

\item{password}{Password for the API}

\item{enterpriseID}{Enterprise ID}
}
\description{
This function queries the Archway API for information on entity classes
}
