% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetIncomeStatement.R
\name{GetIncomeStatement}
\alias{GetIncomeStatement}
\title{GetIncomeStatement API Call}
\usage{
GetIncomeStatement(username, password, enterpriseID, StartDate, EndDate)
}
\arguments{
\item{username}{Username for the API}

\item{password}{Password for the API}

\item{enterpriseID}{Enterprise ID}

\item{StartDate}{Date for beginning of period in yyyy-mm-dd format}

\item{EndDate}{Date for end of period in yyyy-mm-dd format}
}
\description{
This function queries the Archway API for Income Statement data for all Entities.
}
