% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPositions.R
\name{GetPositions}
\alias{GetPositions}
\title{GetPositions API Call}
\usage{
GetPositions(username, password, enterpriseID, StartDate, EndDate)
}
\arguments{
\item{username}{Username for the API}

\item{password}{Password for the API}

\item{enterpriseID}{Enterprise ID}

\item{StartDate}{Date for beginning of period in yyyy-mm-dd format}

\item{EndDate}{Date for end of period in yyyy-mm-dd format}
}
\description{
This function queries the Archway API for information on Open Positions. It is programmed to fetch positions for all entities
}
