% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPositions.R
\name{GetPositionsRaw}
\alias{GetPositionsRaw}
\title{GetPositionsRaw API Call}
\usage{
GetPositionsRaw(username, password, enterpriseID, StartDate, EndDate)
}
\arguments{
\item{username}{Username for the API}

\item{password}{Password for the API}

\item{enterpriseID}{Enterprise ID}

\item{StartDate}{Date for beginning of period in yyyy-mm-dd format}

\item{EndDate}{Date for end of period in yyyy-mm-dd format}
}
\description{
This function queries the Archway API for information on Positions. It is programmed to fetch positions for all entities. This returns XML response
}
