% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSecurities.R
\name{GetSecurityClasses}
\alias{GetSecurityClasses}
\title{GetSecurityClasses API Call}
\usage{
GetSecurityClasses(username, password, enterpriseID)
}
\arguments{
\item{username}{Username for the API}

\item{password}{Password for the API}

\item{enterpriseID}{Enterprise ID}
}
\description{
This function queries the Archway API for information on security classes
}
