% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSecurityPerformance.R
\name{GetSecurityPerformance}
\alias{GetSecurityPerformance}
\title{GetSecurityPerformance API Call}
\usage{
GetSecurityPerformance(
  username,
  password,
  enterpriseID,
  StartDate,
  EndDate,
  Metric = c("TWR", "IRR", "XIRR"),
  Annualized = FALSE
)
}
\arguments{
\item{username}{Username for the API}

\item{password}{Password for the API}

\item{enterpriseID}{Enterprise ID}

\item{StartDate}{Date for beginning of period in yyyy-mm-dd format}

\item{EndDate}{Date for end of period in yyyy-mm-dd format}
}
\description{
This function queries the Archway API for Performance data. It is programmed to fetch performance for all entities
}
