% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSecurityValues.R
\name{GetSecurityValues}
\alias{GetSecurityValues}
\title{GetSecurityValues API Call}
\usage{
GetSecurityValues(
  username,
  password,
  enterpriseID,
  StartDate = "1900-01-01",
  EndDate = Sys.Date()
)
}
\arguments{
\item{username}{Username for the API}

\item{password}{Password for the API}

\item{enterpriseID}{Enterprise ID}

\item{StartDate}{Date for beginning of period in yyyy-mm-dd format}

\item{EndDate}{Date for end of period in yyyy-mm-dd format}
}
\description{
This function queries the Archway API for information from the Security Master for a given period
}
